Remove a NAL unit's trailing zero bytes even when dst_length is 1. Consider the following byte sequence
    00 00 01 0a 00 00 00 01 09 ...
               ^  ^
               A  B

decode_nal() determines dst_length to be 1 (i. e. the byte between label
A and B above). However, this byte is a trailing zero byte as the spec
says the the current NAL unit is terminated by a byte sequence 00 00 00.

The current code used a loop to decrement dst_length accordingly. But the
loop doesn't start as the loop condition checks for dst_length > 1, which
should read dst_length > 0.
patch by Reinhard Nissl, rnissl gmx de


git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@8689 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b