Fix pcm_read_seek () when the position it calculates is greater than 2 GB.
pcm_read_seek() puts the return value of url_fseek() in an int and then
compares < 0 to see if an error occurred; if the position is greater
than 2 GB, the 32-bit signed int result will be < 0.

Change the type of ret to int64_t to avoid the wraparound.

patch by Daniel Verkamp, daniel drv nu


git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@20169 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b