configure: fix symver detection with icc 10
Apparently the compiler sometimes bypasses the normal assembler, in
which case the asm label variant works.  If the file contains any
inline asm, this bypassing is not done, and the assembler chokes on
invalid syntax.  Inserting a dummy asm() statement in the test when
the compiler supports it makes sure no shortcuts are taken and the
test fails as it should.

git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@23616 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b