Fix lavfi pixdesc test
This test verifies the pixdesc code by comparing the output with and
without a filter which should have no effect on the image.  Since the
available pixel formats depend on the byte order of the machine, a
simple reference checksum is not possible.

The test originally tried to solve this by generating a reference file
on the fly.  The problem with this is that the test framework expects
the reference file in the source tree, and writing to the source tree
is not allowed.

To avoid complicating the test framework, we instead provide two
reference files and select which to use based on the byte order.

git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@24330 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b