Make pkgconfig_generate() explicitly return 0 in instead of returning without value when the target library is disabled.
If it does not explicitly return 0, when the last library is disabled
(swscale), the final "configure" exit value is 1, even if the
configure script is successfully executed. So it breaks scripts that
invoke configure and rely on 0 for success and 1 for failure.

Patch by Jean-Daniel Dupas reverse(<org.shadowlab@devlists>).

git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@24585 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b