Make avfilter_start_frame() invoke avfilter_get_video_buffer() on the link rather than avfilter_default_get_video_buffer().
This is required as the buffer requested may be greater than the
buffer allocated locally by avfilter_default_get_video_buffer(), for
example if in filterchain there is a pad filter (like in "fifo,pad").

In that case the pad filter will try to write beyond the data of the
allocated buffer, resulting in a crash.

git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@24895 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b