configure: Set _DARWIN_C_SOURCE while testing for struct ipv6_mreq
On OS X, this struct is only available if _DARWIN_C_SOURCE is defined,
when we have _POSIX_C_SOURCE defined. This allows the struct to be found,
enabling proper IPv6 multicast functionality on OS X.

The define is already set within the file that uses the struct. Setting it
only for this test in configure avoids having to keep it defined for the
whole build.


git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@25370 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b