Replace libswscale svn:external by last revision from the MPlayer repository.
The history of libswscale is now available directly in git, which is preferred
for development, while this repository now serves mainly historic purposes.
Furthermore the libswscale directory in MPlayer is not guaranteed to continue
to exist for as long as this repository stays available, so it makes sense to
keep the last revision in this repository self-contained and forever buildable.


git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk@26402 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b